Выполнены рекомендации ревьюера
