+++ ImageFeed/Auth/OAuth2TokenStorage.swift
@@ -10,6 +10,7 @@ import SwiftKeychainWrapper
 
 final class OAuth2TokenStorage {
     static let shared = OAuth2TokenStorage()
+    private init(){}
     private let keychainStorage = KeychainWrapper.standard
     
     private enum Keys: String {
